+++ src/utils/settings.py
@@ -111,7 +111,7 @@ APPLICATION = "commix"
 DESCRIPTION_FULL = "Automated All-in-One OS Command Injection and Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
-VERSION_NUM = "2.7.5"
+VERSION_NUM = "2.7.6"
 STABLE_VERSION = False
 if STABLE_VERSION:
   VERSION = "v" + VERSION_NUM[:3] + "-stable"
@@ -687,11 +687,12 @@ TAMPER_SCRIPTS = {
                   "space2vtab": False,
                   "singlequotes": False,
                   "caret": False,
-                  "multiplespaces":False,
-                  "backslashes":False,
-                  "nested":False,
-                  "sleep2usleep":False,
-                  "sleep2timeout":False
+                  "multiplespaces": False,
+                  "backslashes": False,
+                  "nested": False,
+                  "sleep2usleep": False,
+                  "sleep2timeout": False,
+                  "xforwardedfor": False
                  }
 
 # HTTP Errors